requirements.txt
flask
flask-cors
firebase-admin
requests
beautifulsoup4
langchain
chromadb
onnxruntime 
sentence-transformers
python-dotenv
langgraph
nltk
gunicorn
langchain-google-genai
google-generativeai
google-cloud-secret-manager
